Now able to have different number of critters shown in train stage versus test stage. This may not be the best way to do this, but there are two arrays (train vs test) for both players.
I started working on a way to make the tests be half on what the player learned originally and half on what the partner learned. To see this look around line 350 in game core. In testing this, I have found it disconnects unexpectantly too. Not sure if this is due to my playing with the code or my local host acting up. Anyway, this part still has a lot of work - I dont know if and when test list is accessed, or if the code I have there ever gets executed.